refactor(observable): move types to folder
