+++ utils/observable/src/types/index.ts
@@ -1,0 +1,6 @@
+export * from './actions';
+export * from './flow';
+export * from './observable';
+export * from './reducer';
+
+export { Flow as Epic } from './flow';
