+++ utils/observable/src/types/observable.ts
@@ -1,0 +1,5 @@
+import { Observable, ObservableInput } from 'rxjs';
+
+export { Observable, ObservableInput };
+
+export type ObservableType<T> = T extends Observable<infer U> ? U : never;
